Рефакторинг utils.js
