+++ src/view/event-view.js
@@ -1,5 +1,5 @@
 import { offersAll } from '../mock/offer.js';
-import { getDuration, getHumanizeDay, getHumanizeTime, getMarkupDate, getMarkupTime } from '../utils.js';
+import { getDuration, getHumanizeDay, getHumanizeTime, getMarkupDate, getMarkupTime } from '../utils/date-time.js';
 import AbstractView from '../framework/view/abstract-view.js';
 
 const createEventTemplate = (event) => {
